Create prepTS_Add_Seconds.py
1 / Issue is that the timestamp (occurrences of events) are NOT evenly/equally spaced, when in general it is assumed to be.

2 / We need a prediction on the amount of data extracted but ALSO on When it is predicted to be extracted

I have not found anything for that. My solution is to convert the timestamp in ISO like (nbr of secs since Jan 1 1970) and make prediction on such a series. The series is now made of 2 variables (amount and nbr of secs) and we should be able to make predictions on both using multi variate ARIMA for example.